+++ tests/files/src/0812_invalid_dim_empty.php
@@ -1,0 +1,14 @@
+<?php
+$x = [];
+var_export([][rand(0,10)]);
+var_export($x[rand(0,10)]);
+var_export($x[0]);
+var_export(''[0]);
+var_export(''[rand(0,1)]);
+var_export('0'[0]);
+var_export('1'[0]);
+var_export('0'[-1]);
+var_export('0'[-2]);
+var_export('0'[1]);
+$x = 'test';
+var_export($x[4]);